+++ src/cli.rs
@@ -58,6 +58,9 @@ pub struct Cli {
     /// Use the tile boundingVolume as the content boundingVolume, instead of calculating the content boundingVolume from the data.
     #[arg(long = "3dtiles-content-bv-from-tile")]
     pub cesium3dtiles_content_bv_from_tile: bool,
+    /// Add the boundingVolume of the content for the the tiles that have content.
+    #[arg(long = "3dtiles-content-add-bv")]
+    pub cesium3dtiles_content_add_bv: bool,
     /// Set the geometric error (see 3D Tiles specification) on the parent nodes of leafs. This controls at what
     /// camera distance leaf nodes become visible. Higher values make content visible earlier when zooming in.
     #[arg(long, short = 'e', default_value = "12")]